Make controllers structure more flat
