+++ src/Http/Controllers/Api/Jobs/CollectAction.php
@@ -11,15 +11,15 @@
 
 declare(strict_types=1);
 
-namespace Cog\Laravel\Paket\Http\Controllers\Api\Jobs\Get;
+namespace Cog\Laravel\Paket\Http\Controllers\Api\Jobs;
 
 use Cog\Contracts\Paket\Job\Repositories\Job as JobRepositoryContract;
 use Illuminate\Contracts\Support\Responsable as ResponsableContract;
 
-final class Action
+final class CollectAction
 {
-    public function __invoke(string $id, JobRepositoryContract $jobs): ResponsableContract
+    public function __invoke(JobRepositoryContract $jobs): ResponsableContract
     {
-        return new Response($jobs->getById($id));
+        return new CollectResponse($jobs->all());
     }
 }
